+++ philibs/scene/img/imgtarga.cpp
@@ -101,7 +101,7 @@ targa* targa::loadHelper ( const std::string& fname )
 			
 			BufferType* buff = new BufferType[ fileStat.st_size ];
 			// TODO: Error check fread.
-			fread ( buff, fileStat.st_size, 1, fid );
+			fread ( buff, ( size_t ) fileStat.st_size, 1, fid );
 			
 			targa* pImg = new targa ( 0 );
